requirements.txt
scikit-learn
pandas
pandas-stubs
scipy
numpy
matplotlib
seaborn
plotly
openpyxl
opencv-python-headless
opencv-python
ultralytics

grpcio
grpcio-tools
boto3
joblib